+++ download_archive_dump.hs
@@ -30,18 +30,8 @@ import Text.Megaparsec
 import Text.Megaparsec.Char (char, digitChar, string)
 import Text.XML.Light (Attr, CData (cdData), Content (Elem, Text), Element (Element, elAttribs, elContent, elName), QName (qName), parseXML)
 
-databaseDumpIndexURL :: URLString
-databaseDumpIndexURL = "https://storage.googleapis.com/mina-archive-dumps/"
-
-curlOptions :: [CurlOption]
-curlOptions = []
-
-getDatabaseDumpIndex :: IO [Content]
-getDatabaseDumpIndex = withCurlDo $ do
-  response :: CurlResponse_ [(String, String)] String <-
-    curlGetResponse_ databaseDumpIndexURL curlOptions
-
-  return . parseXML . respBody $ response
+import Lib.ArchiveDumpKeyParser
+import Lib.Fetchers (fetchDatabaseDumpIndex, fetchArchiveDump)
 
 getListBucketsResult :: Content -> [Content]
 getListBucketsResult = \case
@@ -77,58 +67,15 @@ getDevnetKeys =
           _ -> Nothing
     )
 
-type Parser = Parsec Void String
-
-data DevnetDump = DevnetDump
-  { version :: Int,
-    year :: Int,
-    month :: Int,
-    day :: Int
-  }
-  deriving (Show, Eq)
-
-instance Ord DevnetDump where
-  compare d1 d2
-    | year d1 /= year d2 = compare (year d1) (year d2)
-    | month d1 /= month d2 = compare (month d1) (month d2)
-    | otherwise = compare (day d1) (day d2)
-
-devnetVersion :: Parser Int
-devnetVersion =
-  try (1 <$ string "devnet-archive-dump-")
-    <|> (2 <$ string "devnet2-archive-dump-")
-
-devnetDate :: Parser (Int, Int, Int)
-devnetDate = do
-  year <- many digitChar
-  char '-'
-  month <- many digitChar
-  char '-'
-  day <- many digitChar
-  pure (read year, read month, read day)
-
-devnetParser :: Parser DevnetDump
-devnetParser = do
-  version <- try devnetVersion
-  (year, month, day) <- devnetDate
-  pure $ DevnetDump version year month day
-
-associateKeyMetadata :: [String] -> [(String, DevnetDump)]
+associateKeyMetadata :: [String] -> [(String, ArchiveDumpKey)]
 associateKeyMetadata keys =
   let metadata = rights . map (runParser devnetParser "") $ keys
    in zip keys metadata
 
-getArchiveDump :: String -> IO ByteString
-getArchiveDump key = withCurlDo $ do
-  response :: CurlResponse_ [(String, String)] ByteString <-
-    curlGetResponse_ (databaseDumpIndexURL ++ key) curlOptions
-
-  return (respBody response)
-
 main :: IO ()
 main = do
   putStrLn "getting database backup keys..."
-  devnetKeys <- getDatabaseDumpIndex <&> (getDevnetKeys . getListBucketsResult . (!! 1))
+  devnetKeys <- fetchDatabaseDumpIndex <&> (getDevnetKeys . getListBucketsResult . (!! 1))
   let keysByDate = sortBy (\(_, x) (_, y) -> x `compare` y) $ associateKeyMetadata devnetKeys
   let (targetKey, _) = last keysByDate
   let archiveDumpTar = "database_dumps/" ++ targetKey
@@ -137,7 +84,7 @@ main = do
   archiveDumpExists <- doesFileExist archiveDumpFilename
   when archiveDumpExists $ do
     putStr
-      ( "archive dump \""
+      ( "\narchive dump \""
           ++ archiveDumpFilename
           ++ "\" exists, would you like to overwrite? (y/N): "
       )
@@ -146,7 +93,7 @@ main = do
 
   putChar '\n'
   putStrLn $ "donwloading archive dump..." ++ targetKey
-  archiveDumpCompressed <- getArchiveDump targetKey
+  archiveDumpCompressed <- fetchArchiveDump targetKey
   let archiveDump = decompress . fromStrict $ archiveDumpCompressed
 
   putChar '\n'
